Проектная работа 14.1

+++ controllers/cards.js
@@ -2,41 +2,52 @@ const mongoose = require('mongoose');
 const Card = require('../models/card');
 
 const ERROR_CODE = 400;
-
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка: ${err}` }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        return res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные' });
+        next(new Error('Переданы некорректные данные'));
+      } else {
+        next(err);
       }
-      return res.status(500).send({ message: `Произошла ошибка: ${err}` });
     });
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
+  const userId = req.user._id;
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
     return res.status(ERROR_CODE).send({ message: 'Некорректный id карточки' });
   }
-  return Card.findByIdAndRemove(cardId)
+  return Card.findById(cardId)
     .then((card) => {
       if (!card) {
         return res.status(404).send({ message: 'Карточка не найдена' });
       }
-      return res.send({ data: card });
+      if (card.owner.toString() !== userId) {
+        return res.status(403).send({ message: 'У вас нет прав для удаления этой карточки' });
+      }
+      return Card.findByIdAndRemove(cardId)
+        .then((deletedCard) => {
+          if (!deletedCard) {
+            return res.status(404).send({ message: 'Карточка не найдена' });
+          }
+          return res.send({ data: deletedCard });
+        })
+        .catch(next);
     })
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка: ${err}` }));
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   const { cardId } = req.params;
   const userId = req.user._id;
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
@@ -53,10 +64,10 @@ module.exports.likeCard = (req, res) => {
       }
       return res.send({ data: card });
     })
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка: ${err}` }));
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   const { cardId } = req.params;
   const userId = req.user._id;
   if (!mongoose.Types.ObjectId.isValid(cardId)) {
@@ -73,5 +84,5 @@ module.exports.dislikeCard = (req, res) => {
       }
       return res.send({ data: card });
     })
-    .catch((err) => res.status(500).send({ message: `Произошла ошибка: ${err}` }));
+    .catch(next);
 };
